pyproject.toml
[tool.poetry]
name = "doctrpp"
version = "0.1.0"
description = "Document Image Processing with PaddlePaddle"
authors = ["Yale Huang <yale@hwang.men>"]
readme = "README.md"
packages = [
    { include = "doctrpp" }
]

[tool.poetry.dependencies]
python = "^3.10"
hdf5storage = "^0.1.19"
loguru = "^0.7.3"
numpy = "^2.2.3"
scipy = "^1.15.2"
opencv-python = "^4.11.0.86"
matplotlib = "^3.10.1"
albumentations = "^2.0.4"
paddlepaddle-gpu = "^2.6.2"

[tool.poetry.scripts]
doctrpp = "doctrpp.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
